Add support for more chunks per frame
This issue appeared for first time here:
https://community.aseprite.org/t/1762/4

+++ src/dio/aseprite_common.h
@@ -68,7 +68,7 @@ struct AsepriteHeader {
 struct AsepriteFrameHeader {
   uint32_t size;
   uint16_t magic;
-  uint16_t chunks;
+  uint32_t chunks;
   uint16_t duration;
 };
